rename AddressRegistry -> address-registry for consistency in file names
